Refactor: Redesign ReminderListView to MVP
